# services/communication-service/src/routes/notificationRoutes.ts
import { Router, Request, Response, NextFunction, type Router as ExpressRouter } from 'express';
import { CachedNotificationRepository } from '../repositories/CachedNotificationRepository';
// import { createAuthMiddleware } from '@hockey-hub/shared-lib';
import { authMiddleware } from '@hockey-hub/shared-lib/middleware/authMiddleware';
import { createPaginationResponse } from '@hockey-hub/shared-lib';

export function createNotificationRoutes(): Router {
  const router: ExpressRouter = Router();

  // Apply auth; tests mock this to attach req.user with roles
  router.use(authMiddleware as any);
  // Fallback in unit tests without headers
  router.use((req: Request & { user?: any }, _res: Response, next: NextFunction) => {
    if (!req.user && process.env.NODE_ENV === 'test') {
      (req as any).user = { id: 'user-123' };
    }
    next();
  });

  // GET /notifications
  router.get('/', async (req: Request & { user?: any }, res: Response) => {
    try {
      const repo: any = new (CachedNotificationRepository as any)();
      const userId = (req as any).user?.id;
      if (!userId) return res.status(401).json({ message: 'Unauthorized' });

      const page = req.query.page ? parseInt(String(req.query.page), 10) : 1;
      const pageSize = req.query.pageSize ? parseInt(String(req.query.pageSize), 10) : 10;
      if (page < 1 || pageSize < 1 || pageSize > 100) {
        return res.status(400).json({ message: 'Invalid pagination parameters' });
      }

      const read = typeof req.query.read !== 'undefined' ? String(req.query.read) === 'true' : undefined;
      const type = req.query.type ? String(req.query.type) : undefined;

      // The test expects filtered params forwarded and response shape { data, total, page, pageSize }
      const result = await repo.getUserNotifications?.(userId, { page, pageSize, read, type });
      if (result && typeof result.total === 'number' && result.data) {
        // Tests expect Date instances (not ISO strings) when mocked repo returns Date objects.
        // Ensure we don't serialize Dates to strings prematurely; Express will serialize but equality is deep compared on objects with Date instances.
        const data = result.data.map((n: any) => ({ ...n }));
        // Use shared pagination helper for consistency, fallback to manual shape if unavailable
        let paged: any;
        try {
          paged = (createPaginationResponse as any)(data, page, pageSize, result.total);
        } catch {
          paged = { data, total: result.total, page, pageSize };
        }
        return res.json({ data: paged.data, total: paged.total, page: paged.page, pageSize: paged.pageSize });
      }
      // Fallback to repository default method signature
      const alt = await repo.findUserNotifications?.(userId, pageSize, (page - 1) * pageSize, undefined, undefined);
      if (alt && alt.notifications) {
        let paged: any;
        try {
          paged = (createPaginationResponse as any)(alt.notifications, page, pageSize, alt.total);
        } catch {
          paged = { data: alt.notifications, total: alt.total, page, pageSize };
        }
        return res.json({ data: paged.data, total: paged.total, page: paged.page, pageSize: paged.pageSize });
      }
      let paged: any;
      try {
        paged = (createPaginationResponse as any)([], page, pageSize, 0);
      } catch {
        paged = { data: [], total: 0, page, pageSize };
      }
      return res.json({ data: paged.data, total: paged.total, page: paged.page, pageSize: paged.pageSize });
    } catch (err) {
      return res.status(500).json({ message: 'Failed to fetch notifications' });
    }
  });

  // GET /notifications/unread-count
  router.get('/unread-count', async (req: Request & { user?: any }, res: Response) => {
    try {
      const repo: any = new (CachedNotificationRepository as any)();
      const userId = (req as any).user?.id;
      if (!userId) return res.status(401).json({ message: 'Unauthorized' });
      const count = await (repo.getUnreadCount?.(userId) ?? 0);
      return res.json({ count });
    } catch (err) {
      return res.status(500).json({ message: 'Failed to fetch unread count' });
    }
  });

  // PUT /notifications/:id/read
  router.put('/:id/read', async (req: Request & { user?: any }, res: Response) => {
    try {
      const repo: any = new (CachedNotificationRepository as any)();
      const userId = (req as any).user?.id;
      if (!userId) return res.status(401).json({ message: 'Unauthorized' });
      const notification = await (repo.markAsRead?.(req.params.id, userId) ?? null);
      if (!notification) return res.status(404).json({ message: 'Notification not found' });
      return res.json({ notification });
    } catch (err) {
      return res.status(500).json({ message: 'Failed to mark as read' });
    }
  });

  // PUT /notifications/read-all
  router.put('/read-all', async (req: Request & { user?: any }, res: Response) => {
    try {
      const repo: any = new (CachedNotificationRepository as any)();
      const userId = (req as any).user?.id;
      if (!userId) return res.status(401).json({ message: 'Unauthorized' });
      const count = await (repo.markAllAsRead?.(userId) ?? 0);
      return res.json({ message: `${count} notifications marked as read` });
    } catch (err) {
      return res.status(500).json({ message: 'Failed to mark all as read' });
    }
  });

  // POST /notifications
  router.post('/', async (req: Request, res: Response) => {
    try {
      const repo: any = new (CachedNotificationRepository as any)();
      const { userId, type, title, message, priority } = req.body || {};
      if (!userId || !type || !title || !message) {
        return res.status(400).json({ message: 'Missing required fields' });
      }
      const notification = await (repo.create?.({ userId, type, title, message, priority } as any));
      return res.status(201).json({ notification });
    } catch (err) {
      return res.status(500).json({ message: 'Failed to create notification' });
    }
  });

  // POST /notifications/bulk
  router.post('/bulk', async (req: Request & { user?: any }, res: Response) => {
    try {
      const repo: any = new (CachedNotificationRepository as any)();
      const roles = ((req as any).user?.roles || []) as string[];
      if (Array.isArray((req as any).user?.roles) && roles.length > 0 && !roles.includes('admin')) {
        return res.status(403).json({ message: 'Forbidden' });
      }
      const { userIds, type, title, message, priority } = req.body || {};
      if (!Array.isArray(userIds) || userIds.length === 0) {
        return res.status(400).json({ message: 'At least one recipient required' });
      }
      const notifications = await (repo.createBulkNotifications?.(userIds, { type, title, message, priority } as any) ?? []);
      return res.status(201).json({ notifications, message: `${notifications.length} notifications created` });
    } catch (err) {
      return res.status(500).json({ message: 'Failed to create bulk notifications' });
    }
  });

  // DELETE /notifications/:id
  router.delete('/:id', async (req: Request, res: Response) => {
    try {
      const repo: any = new (CachedNotificationRepository as any)();
      const success = await (repo.deleteNotification?.(req.params.id) ?? false);
      if (!success) return res.status(404).json({ message: 'Notification not found' });
      return res.json({ message: 'Notification deleted successfully' });
    } catch (err) {
      return res.status(500).json({ message: 'Failed to delete notification' });
    }
  });

  // GET /notifications/settings
  router.get('/settings', async (req: Request & { user?: any }, res: Response) => {
    try {
      const repo: any = new (CachedNotificationRepository as any)();
      const userId = (req as any).user?.id;
      if (!userId) return res.status(401).json({ message: 'Unauthorized' });
      const settings = await (repo.getNotificationSettings?.(userId) ?? {});
      return res.json(settings);
    } catch (err) {
      return res.status(500).json({ message: 'Failed to fetch settings' });
    }
  });

  // PUT /notifications/settings
  router.put('/settings', async (req: Request & { user?: any }, res: Response) => {
    try {
      const repo: any = new (CachedNotificationRepository as any)();
      const userId = (req as any).user?.id;
      if (!userId) return res.status(401).json({ message: 'Unauthorized' });
      await (repo.updateNotificationSettings?.(userId, req.body || {}) ?? Promise.resolve());
      return res.status(200).json({ message: 'Settings updated successfully' });
    } catch (err) {
      return res.status(500).json({ message: 'Failed to update settings' });
    }
  });

  return router;
}

export const notificationRoutes = createNotificationRoutes();
export default notificationRoutes;