# services/communication-service/src/swagger/communication-service.yaml
openapi: 3.0.3
info:
  title: Hockey Hub Communication Service API
  description: |
    The Communication Service manages messaging, notifications, chat, and communication features for the Hockey Hub platform.
    This service handles real-time chat, notifications, email/SMS delivery, and communication analytics.
  version: 1.0.0
  contact:
    name: Hockey Hub Development Team
    email: dev@hockeyhub.com
servers:
  - url: http://localhost:3002
    description: Development server
  - url: https://api.hockeyhub.com/communication
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /notifications:
    get:
      tags:
        - Notifications
      summary: Get user notifications
      description: Get paginated notifications for the authenticated user
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: type
          in: query
          schema:
            type: string
            enum: [message, system, training, medical, payment]
        - name: status
          in: query
          schema:
            type: string
            enum: [unread, read, archived]
      responses:
        200:
          description: Notifications retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Notification'

    post:
      tags:
        - Notifications
      summary: Create notification
      description: Create a new notification (system/admin use)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotificationRequest'
      responses:
        201:
          description: Notification created successfully
  /notifications/{id}/read:
    put:
      tags:
        - Notifications
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Notification marked as read
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Notification:
      type: object
      properties:
        id:
          type: string
          example: "notif_123"
        userId:
          type: string
        type:
          type: string
          enum: [message, system, training, medical, payment]
        title:
          type: string
        content:
          type: string
        data:
          type: object
        status:
          type: string
          enum: [unread, read, archived]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        channels:
          type: array
          items:
            type: string
            enum: [push, email, sms, in_app]
        scheduledFor:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    CreateNotificationRequest:
      type: object
      properties:
        userIds:
          type: array
          items:
            type: string
        type:
          type: string
          enum: [message, system, training, medical, payment]
        title:
          type: string
        content:
          type: string
        data:
          type: object
        priority:
          type: string
          enum: [low, normal, high, urgent]
        channels:
          type: array
          items:
            type: string
            enum: [push, email, sms, in_app]
        scheduledFor:
          type: string
          format: date-time
      required:
        - userIds
        - type
        - title
        - content
...